Change the method for detecting temperature
Use counter to indicates that the temperature is in or out of the range,
if the counter reach certain amount then change the status of INMS
handler.
In this way, it will solve the issue if temperature is changing between
the threshold, causing the status keep changing all the time.